# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger Yummi - OpenAPI 3.0",
        "description": "Swagger documentation for project So Yummi",

        "servers": [
            {
                "url": "https://team-project-so-yummy-backend.onrender.com/api/"
            }
        ],
        "tags": [
            { "name": "auth", "description": "Authorization endpoints" },
            { "name": "recipes", "description": "Recipes endpoints" }
        ]
    },
    "paths": {
...
        "/api/auth/verify/{verificationToken}": {
            "get": {
                "tags": ["auth"],
                "summary": "Email verify",
                "description": "Email verify",
                "operationId": "verifyEmail",
                "parameters": [
                    {
                        "in": "path",
                        "name": "verificationToken",
                        "required": true,
                        "type": "string",
                        "description": "verification token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification successful"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        },
...
    },
...
}